% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aphia.R
\docType{data}
\name{aphia}
\alias{aphia}
\title{WoRMS AphiaIDs for ICES assessment species.}
\format{A data frame with 319 rows and 2 variables:
\describe{
  \item{aphia_code}{numeric identifier code as used in WoRMS database}
  \item{species}{Latin species name, e.g. Myxine glutinosa}
}}
\source{
\url{http://vocab.ices.dk/}

\url{http://www.marinespecies.org/aphia.php}
}
\usage{
aphia
}
\description{
A dataset containing the aphia codes and latin species names
of species assessed by ICES.
}
\keyword{datasets}

