% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputs.R
\name{checkSurveyOK}
\alias{checkSurveyOK}
\title{Check that a survey name is in the database}
\usage{
checkSurveyOK(survey)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}
}
\value{
logical.
}
\description{
Checks a survey name against a list of all survey names in the DATRAS database.
If the name is not matched it puts up a message showing the available survey names.
}
\examples{
\dontrun{
checkSurveyOK(survey = "ROCKALL")
checkSurveyOK(survey = "NOTALL")
}
}
\seealso{
\code{\link{checkSurveyYearOK}} and \code{\link{getSurveyYearQuarterList}}
 also perform checks against the DATRAS database.

\code{\link{icesDatras-package}} gives an overview of the package.
}
