% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchWgt.R
\name{getCatchWgt}
\alias{getCatchWgt}
\title{Get Reported Catch Weights}
\usage{
getCatchWgt(survey, years, quarters, aphia_codes)
}
\arguments{
\item{survey}{the survey acronym e.g. NS-IBTS.}

\item{years}{a vector of numeric years of the survey, e.g. c(2010, 2012), or 2005:2010.}

\item{quarters}{a vector of quarters of the year the survey took place, i.e. c(1, 4) or 1:4.}

\item{aphia_codes}{a vector of numeric 'aphia' codes defined in the WoRMS database, i.e. c(1, 4) or 1:4.}
}
\value{
A data frame.
}
\description{
Get the reported catch weight by species and haul.
}
\examples{
getCatchWgt(survey = "ROCKALL", years = 2002, quarters = 3, aphia_codes = "126437")

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyYearList}}, \code{\link{getSurveyYearQuarterList}}, and
\code{\link{getDatrasDataOverview}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}

