% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPUEAge.R
\name{getCPUEAge}
\alias{getCPUEAge}
\title{Get CPUE per Age per Haul per Hour for a given Survey, year and Quarter}
\usage{
getCPUEAge(survey, year, quarter)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{year}{the year of the survey, e.g. 2010.}

\item{quarter}{the quarter of interest, e.g. 1}
}
\value{
A data frame.
}
\description{
Get CPUE per Age per Haul per Hour for a given Survey, year and Quarter
}
\examples{
\dontrun{
getCPUEperAge(survey = "NS-IBTS", year = 2018, quarter = 1)
}
}
\seealso{
\code{\link{getHHdata}} and \code{\link{getCAdata}} get haul data and
age-based data., and
\code{\link{getIndices}}, and
\code{\link{getCPUELength}}.

\code{\link{icesDatras-package}} gives an overview of the package.
}
\author{
Adriana Villamor.
}
