% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatrasDataOverview.R
\name{getDatrasDataOverview}
\alias{getDatrasDataOverview}
\title{Summarize Data Availability}
\usage{
getDatrasDataOverview()
}
\value{
A list of tables.
}
\description{
Returns a table for each survey with '1' where there is data and '0' (printed as '.') otherwise.
}
\examples{
\dontrun{
getDatrasDataOverview()
}

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyList}} returns the acronyms for available surveys.

\code{\link{getSurveyYearList}} returns the years available for a given survey.

\code{\link{getSurveyYearQuarterList}} returns the quarters available for a given survey and year.

\code{\link{icesDatras-package}} gives an overview of the package.
}

