% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatrasFieldList.R
\name{getDatrasFieldList}
\alias{getDatrasFieldList}
\title{Get Datras field list with column classes and mapping to old naming schema}
\usage{
getDatrasFieldList()
}
\value{
A dataframe
}
\description{
Get Datras field list with column classes and mapping to old naming schema
}
