% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLTassessment.R
\name{getLTassessment}
\alias{getLTassessment}
\title{Get Litter assessment output}
\usage{
getLTassessment(
  survey,
  year,
  quarter,
  fix_types = getOption("icesDatras.fix_types"),
  new_names = getOption("icesDatras.new_names")
)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{year}{the year of the survey, e.g. 2010.}

\item{quarter}{the quarter of the year the survey took place, i.e. 1, 2, 3 or 4.}

\item{fix_types}{logical, apply the DATRAS type to columns. Takes package default 
unless specified. Use \code{SetDatrasDefaults()} to change 
default across all functions.}

\item{new_names}{logical, apply the new DATRAS naming convention to output. 
Takes package default unless specified. Use 
\code{SetDatrasDefaults()} to change default across all functions.}
}
\value{
A data frame.
}
\description{
Get Litter assessment output by survey, year and quarter. The raw data are also
included in this file.
}
\note{
The \pkg{icesAdvice} package provides \code{findAphia}, a function to look up Aphia species codes.
}
\examples{
\dontrun{
index <- getLTassessment(survey = "NS-IBTS", year = 2002, quarter = 3)
str(index)
}
}
\seealso{
\code{\link{getDATRAS}} supports querying many years and quarters in one function call.

\code{\link{getHHdata}} and \code{\link{getCAdata}} get haul data and
age-based data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
