% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyList.R
\name{getSurveyList}
\alias{getSurveyList}
\title{Query data availability}
\usage{
getSurveyList()
}
\value{
A numeric vector.
}
\description{
Returns list of survey acronyms.
}
\examples{
getSurveyList()


}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyYearList}} returns the years available for a given survey.

\code{\link{getSurveyYearQuarterList}} returns the quarters available for a given survey and year.

\code{\link{icesDatras-package}} gives an overview of the package.
}

