% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyYearList.R
\name{getSurveyYearList}
\alias{getSurveyYearList}
\title{Query data availability}
\usage{
getSurveyYearList(survey)
}
\arguments{
\item{survey}{the survey accronym e.g. NS-IBTS, BITS.}
}
\value{
A numeric vector.
}
\description{
Returns list of years based on survey.
}
\examples{
getSurveyYearList(survey = "NS-IBTS")


}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyList}} returns the acronyms for available surveys.

\code{\link{getSurveyYearQuarterList}} returns the quarters available for a given survey and year.

\code{\link{icesDatras-package}} gives an overview of the package.
}

