% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyYearQuarterList.R
\name{getSurveyYearQuarterList}
\alias{getSurveyYearQuarterList}
\title{Query data availability}
\usage{
getSurveyYearQuarterList(survey, year)
}
\arguments{
\item{survey}{the survey accronym e.g. NS-IBTS, BITS.}

\item{year}{the numeric year of the survey, e.g. 2010.}
}
\value{
A numeric vector.
}
\description{
Returns list of quarter based on survey and year.
}
\examples{
getSurveyYearQuarterList(survey = "NS-IBTS", year = 2010)


}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyList}} returns the acronyms for available surveys.

\code{\link{getSurveyYearList}} returns the years available for a given survey.

\code{\link{icesDatras-package}} gives an overview of the package.
}

