% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesDatras-package.R
\docType{package}
\name{icesDatras-package}
\alias{icesDatras}
\alias{icesDatras-package}
\title{DATRAS trawl database webservices}
\description{
Functions that access the webservices of the ICES (International
Council for the Exploration of the Sea) DATRAS trawl database.
}
\author{
Colin Millar.
}
\references{
Description of ICES Datras web services:
\url{https://datras.ices.dk/WebServices/Webservices.aspx}.

ICES list of software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.

}

