% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDATRAS.R
\name{getDATRAS}
\alias{getDATRAS}
\title{Get Any DATRAS Data}
\usage{
getDATRAS(record = "HH", survey, years, quarters)
}
\arguments{
\item{record}{the data type required: "HH" haul meta data, "HL" length-based data, "CA" age-based data}

\item{survey}{the survey accronym e.g. NS-IBTS, BITS.}

\item{years}{a vector of numeric years of the survey, e.g. c(2010, 2012), or 2005:2010.}

\item{quarters}{a vector of quarters of the year the survey took place, i.e. c(1, 4) or 1:4.}
}
\value{
A data frame.
}
\description{
Get any DATRAS data. This function combines the functionality of getCAdata, getHHdata, and getHLdata.
}
\examples{
\dontrun{
 getDATRAS(record = "HH",
           survey = "NS-IBTS",
           years = 2011:2015,
           quarters = c(1,4))
}

}
\author{
Scott Large and Colin Millar
}

