% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatrasDataOverview.R
\name{getDatrasDataOverview}
\alias{getDatrasDataOverview}
\title{Summarize Data Availability}
\usage{
getDatrasDataOverview()
}
\value{
A list of tables.
}
\description{
Evaluate a presence-absence table for each survey with '1' where there is data and '0' (printed as '.') otherwise.
}
\examples{
\dontrun{
getDatrasDataOverview()
}

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getSurveyList}}, \code{\link{getSurveyYearList}}, and
\code{\link{getSurveyYearQuarterList}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}

