% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatrasDataOverview.R
\name{getDatrasDataOverview}
\alias{getDatrasDataOverview}
\title{Summarize Data Availability}
\usage{
getDatrasDataOverview(surveys = NULL, long = TRUE)
}
\arguments{
\item{surveys}{a vector of survey names, or \code{NULL} to process all surveys.}

\item{long}{whether tables should have year as row names (default) or column names.}
}
\value{
A list of tables.
}
\description{
Evaluate a presence-absence table for each survey with '1' where there is data and '0' (printed as '.') otherwise.
}
\examples{
\dontrun{
getDatrasDataOverview(surveys = "ROCKALL", long = FALSE)
}
}
\seealso{
\code{\link{getSurveyList}}, \code{\link{getSurveyYearList}}, and
\code{\link{getSurveyYearQuarterList}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
