% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatrasUnaggregated.R
\name{getDatrasUnaggregated}
\alias{getDatrasUnaggregated}
\title{Download unaggregated DATRAS survey data}
\usage{
getDatrasUnaggregated(recordtype, survey, year, quarter)
}
\arguments{
\item{recordtype}{Character. One of `"HH"`, `"HL"`, or `"CA"`.}

\item{survey}{Character. Survey acronym (e.g. `"NS-IBTS"`).}

\item{year}{Character. Year or range (e.g. `"2020"` or `"1965:2025"`).}

\item{quarter}{Character. Quarter or range (e.g. `"1"` or `"1:4"`).}
}
\value{
A `data.table` containing the requested DATRAS data.
}
\description{
Downloads unaggregated haul- and biological-level data from the
ICES DATRAS Download API.
}
\details{
The function downloads a zipped CSV file from the official ICES DATRAS API,
extracts it locally, and reads it using fixed column classes to avoid
costly type guessing.
}
\examples{
\dontrun{
df <- get_datras_unaggregated_data(
  recordtype = "HH",
  survey = "NS-IBTS",
  year = "1965:2025",
  quarter = "1:4"
)
}

}
\author{
Vaishav Soni, International Council for the Exploration of the Sea (ICES)
}
