% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHHdata.R
\name{getHHdata}
\alias{getHHdata}
\title{Get Haul Data}
\usage{
getHHdata(survey, year, quarter)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{year}{the year of the survey, e.g. 2010.}

\item{quarter}{the quarter of the year the survey took place, i.e. 1, 2, 3 or 4.}
}
\value{
A data frame.
}
\description{
Get haul data such as position, depth, sampling method, etc.
}
\examples{
# read haul data
hhdata <- getHHdata(survey = "NS-IBTS", year = 2016, quarter = 1)
str(hhdata)

# error checking examples:
hhdata <- getHHdata(survey = "NS_IBTS", year = 2016, quarter = 1)
hhdata <- getHHdata(survey = "NS-IBTS", year = 2018, quarter = 1)
hhdata <- getHHdata(survey = "NS-IBTS", year = 2016, quarter = 6)

# multiple year example
hhdata <- do.call(rbind,
                  lapply(2015:2016,
                         function(year)
                           getHHdata(survey = "NS-IBTS", year = year, quarter = 1)
                         )
                 )
str(hhdata)

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getHLdata}} and \code{\link{getCAdata}} get length-based data and
age-based data.

\code{\link{icesDatras-package}} gives an overview of the package.
}

