% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHLdata.R
\name{getHLdata}
\alias{getHLdata}
\title{Get Length-Based Data}
\usage{
getHLdata(survey, year, quarter)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{year}{the numeric year of the survey, e.g. 2010.}

\item{quarter}{the quarter of the year the survey took place, i.e. 1, 2, 3 or 4.}
}
\value{
A data frame.
}
\description{
Get length-based information such as measured length, individual counts,
  and subfactors of sampled species.
}
\examples{
# read length data
hldata <- getHLdata(survey = "NS-IBTS", year = 2016, quarter = 1)
str(hldata)

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getHHdata}} and \code{\link{getCAdata}} get haul data and
age-based data.

\code{\link{getSurveyList}}, \code{\link{getSurveyYearList}},
\code{\link{getSurveyYearQuarterList}}, and
\code{\link{getDatrasDataOverview}} list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
\keyword{distribution}

