% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyList.R
\name{getSurveyList}
\alias{getSurveyList}
\title{Get a List of All Surveys}
\usage{
getSurveyList()
}
\value{
A character vector.
}
\description{
Get a list of all survey acronyms.
}
\examples{
\dontrun{
getSurveyList()
}
}
\seealso{
\code{\link{getSurveyYearList}}, \code{\link{getSurveyYearQuarterList}}, and
\code{\link{getDatrasDataOverview}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
