% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyYearList.R
\name{getSurveyYearList}
\alias{getSurveyYearList}
\title{Get a List of Survey Years}
\usage{
getSurveyYearList(survey)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}
}
\value{
A numeric vector.
}
\description{
Get a list of all data years for a given survey.
}
\examples{
\dontrun{
getSurveyYearList(survey = "ROCKALL")
}
}
\seealso{
\code{\link{getSurveyList}}, \code{\link{getSurveyYearQuarterList}}, and
\code{\link{getDatrasDataOverview}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
