% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurveyYearQuarterList.R
\name{getSurveyYearQuarterList}
\alias{getSurveyYearQuarterList}
\title{Get a List of Quarters}
\usage{
getSurveyYearQuarterList(survey, year)
}
\arguments{
\item{survey}{the survey acronym, e.g. NS-IBTS.}

\item{year}{the year of the survey, e.g. 2010.}
}
\value{
A numeric vector.
}
\description{
Get a list of quarters available for a given survey and year.
}
\examples{
\dontrun{
getSurveyYearQuarterList(survey = "ROCKALL", year = 2002)
}
}
\seealso{
\code{\link{getSurveyYearList}}, \code{\link{getSurveyYearQuarterList}}, and
\code{\link{getDatrasDataOverview}} also list available data.

\code{\link{icesDatras-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
