% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesDatras-package.R
\docType{package}
\name{icesDatras-package}
\alias{icesDatras-package}
\alias{icesDatras}
\title{DATRAS Trawl Survey Database Web Services}
\description{
R interface to access the web services of the ICES DATRAS trawl survey
database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getHHdata}} \tab haul data\cr
  \code{\link{getHLdata}} \tab length-based data\cr
  \code{\link{getCAdata}} \tab age-based data\cr
  \code{\link{getDATRAS}} \tab any data
}
\emph{Overview of available data:}
\tabular{ll}{
  \code{\link{getSurveyList}}            \tab surveys\cr
  \code{\link{getSurveyYearList}}        \tab years\cr
  \code{\link{getSurveyYearQuarterList}} \tab quarters\cr
  \code{\link{getDatrasDataOverview}}    \tab surveys, years, and quarters
}
\emph{Calculate:}
\tabular{ll}{
  \code{\link{getSurveyList}}            \tab surveys\cr
  \code{\link{getSurveyYearList}}        \tab years\cr
  \code{\link{getSurveyYearQuarterList}} \tab quarters\cr
  \code{\link{getDatrasDataOverview}}    \tab surveys, years, and quarters
}
}
\references{
ICES DATRAS database: \url{http://datras.ices.dk}.

ICES DATRAS web services:
\url{https://datras.ices.dk/WebServices/Webservices.aspx}.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
